+++ src/modules/layerTree/tests/unit/components/LayerCheckBox.spec.js
@@ -202,6 +202,7 @@ describe("src/modules/layerTree/components/LayerCheckBox.vue", () => {
     });
 
     it("renders layer with visibility true and checkbox, name is bold", () => {
+
         propsData.conf.visibility = true;
 
         wrapper = shallowMount(LayerCheckBox, {
@@ -211,9 +212,12 @@ describe("src/modules/layerTree/components/LayerCheckBox.vue", () => {
             propsData
         });
 
-        expect(wrapper.find("#layer-checkbox-" + propsData.conf.id).exists()).to.be.true;
-        expect(wrapper.find("#layer-checkbox-" + propsData.conf.id).attributes("title")).to.not.be.undefined;
-        expect(wrapper.find("#layer-checkbox-" + propsData.conf.id).attributes("aria-checked")).to.equal("true");
+        const checkbox = wrapper.find("#layer-checkbox-" + propsData.conf.id);
+
+        expect(checkbox.exists()).to.be.true;
+        expect(checkbox.attributes("role")).to.equal("checkbox");
+        expect(checkbox.attributes("aria-label")).to.equal(propsData.conf.name);
+        expect(checkbox.attributes("aria-checked")).to.equal("true");
         expect(wrapper.findAll(".layer-tree-layer-checkbox").length).to.be.equals(1);
         expect(wrapper.find(".bi-check-square").exists()).to.be.true;
         expect(wrapper.find(".layer-tree-layer-label").text()).to.equal(propsData.conf.name);